requirements-dev.txt
fastapi==0.109.2
uvicorn==0.27.1
pre-commit==3.6.1
python-multipart==0.0.9
pydantic-settings==2.1.0
minio==7.2.4
pytest-asyncio==0.23.5
aio-pika==9.4.0
